% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maven.R
\name{as.coordinates}
\alias{as.coordinates}
\title{Maven coordinates}
\usage{
as.coordinates(groupId, artifactId, version, ...)
}
\arguments{
\item{groupId}{the maven groupId}

\item{artifactId}{the maven artifactId}

\item{version}{the maven version}

\item{...}{other params ignored apart from packaging (jar or ejb) and classifier (tests, client, sources, javadoc)}
}
\value{
a coordinates object containing the coordinates
}
\description{
Maven coordinates
}
\examples{
as.coordinates("org.junit.jupiter","junit-jupiter-api","4.13.2")
}
