% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maven.R
\name{clear_rmaven_cache}
\alias{clear_rmaven_cache}
\title{Clear out the \code{rmaven} cache}
\usage{
clear_rmaven_cache()
}
\value{
nothing, called for side effects
}
\description{
Deletes all content in the \code{rmaven} cache. This should not be necessary, but never
say never, and if there is really a problem with the cache, then deleting it may be the
best thing. This will wait for confirmation from the user. If running unattended the
\code{options("rmaven.allow.cache.delete"=TRUE)} must be set for the action to occur, otherwise
it will generate a warning and do nothing.
}
\examples{
\donttest{
# need to set the following option to allow cache to be deleted in non interactive session
opts = options("rmaven.allow.cache.delete"=TRUE)
clear_rmaven_cache()
options(opts)
}
}
