% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-rmaven.R
\name{.resolve_dependencies}
\alias{.resolve_dependencies}
\title{Resolve dependencies and calculate the \code{classpath} for an artifact.}
\usage{
.resolve_dependencies(
  groupId = NULL,
  artifactId = NULL,
  version = NULL,
  ...,
  coordinates = NULL,
  artifact = NULL,
  path = NULL,
  include_self = NULL,
  nocache = FALSE,
  verbose = c("normal", "quiet", "debug")
)
}
\arguments{
\item{groupId}{the maven \code{groupId}, optional}

\item{artifactId}{the maven \code{artifactId}, optional}

\item{version}{the maven version, optional}

\item{...}{passed on to as.coordinates()}

\item{coordinates}{the maven coordinates, optional (either
\code{groupId},\code{artifactId} and 'version' must be specified, or 'coordinates',
or 'artifact')}

\item{artifact}{optional, coordinates as an artifact string
\verb{groupId:artifactId:version[:packaging[:classifier]]} string}

\item{path}{the path to the source directory, pom file or jar file. if not
given \code{rmaven} will get the artifact from the maven central repositories}

\item{include_self}{do you want include this path in the \code{classpath}.
optional, if missing the path will be included if it is a regular jar, or a
fat jar, otherwise not.}

\item{nocache}{do not used cached version, by default we use a cached version
of the \code{classpath} unless the \code{pom.xml} is newer that the cached
\code{classpath}.}

\item{verbose}{how much output from maven, one of "normal", "quiet", "debug"}
}
\value{
a character vector of the \code{classpath} jar files (including the
current one if appropriate)
}
\description{
This function makes sure the transitive dependencies for a maven artifact are
available locally in the \code{.m2} maven cache and calculates a local classpath
which can be provided to \code{rJava}. The artifact may be specified either as a
set of maven coordinates (in which case the artifact itself is also
downloaded, and included in the \code{classpath}) or as a path to a jar file
containing a pom.xml (e.g. a compiled jar file, a compiled
\code{...-jar-with-dependencies}, or a assembled \code{...-src.jar}).
}
\keyword{internal}
