% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-rmaven.R
\name{.set_repository_location}
\alias{.set_repository_location}
\title{Sets the local maven repository location}
\usage{
.set_repository_location(
  repository_location = getOption("rmaven.m2.repository", default = .working_dir(subpath
    = ".m2/repository/")),
  settings_path = .settings_path()
)
}
\arguments{
\item{repository_location}{a file path (which will be expanded to a full
path) where the repository should be based, e.g. \verb{~/.m2/repository/}.
Defaults to a sub-directory of the \code{rmaven} cache.}

\item{settings_path}{the file path of the settings.xml to update (generally
the supplied default is what you want to use)}
}
\value{
the expanded path of the new repository location
}
\description{
This writes a maven repository location to a temporary \code{settings.xml} file
which persists only for the R session. The location of the maven repository
is either specified here, or can be defined by the
\code{options("rmaven.m2.repository"=...)} option. If neither of these is
provided, the location will revert to a default location within the \code{rmaven}
cache. (Approved by CRAN for a local cache location) e.g. on 'Linux' this
will default to \verb{~/.cache/rmaven/.m2/repository/}
}
\keyword{internal}
