% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maven.R
\name{execute_maven}
\alias{execute_maven}
\title{Executes a maven goal}
\usage{
execute_maven(
  goal,
  opts = c(),
  pom_path = NULL,
  quiet = .quietly(verbose),
  debug = .debug(verbose),
  verbose = c("normal", "debug", "quiet"),
  require_jdk = FALSE,
  ...
)
}
\arguments{
\item{goal}{the goal of the mvn command ( can be multiple ) e.g. c("clean","compile")}

\item{opts}{provided options in the form c("-Doption1=value2","-Doption2=value2")}

\item{pom_path}{optional. the path to a pom.xml file for goals that operate on one}

\item{quiet}{should output from maven be suppressed? (-q flag)}

\item{debug}{should output from maven be verbose? (-X flag)}

\item{verbose}{how much output from maven, one of "normal", "quiet", "debug"}

\item{require_jdk}{does the goal you are executing require a jdk (e.g. compilation)}

\item{...}{named parameters are passed to maven as options in the form -Dname=value}
}
\value{
nothing, invisibly
}
\description{
Maven goals may be executed with or without a pom.xml file. Some maven goals (e.g. compilation)
require the use of a JDK. Goals can be
}
\examples{
execute_maven("help:system")
}
