% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maven.R
\name{fetch_artifact}
\alias{fetch_artifact}
\title{Fetch an artifact from a repository into the local .m2 cache}
\usage{
fetch_artifact(
  groupId = NULL,
  artifactId = NULL,
  version = NULL,
  ...,
  coordinates = NULL,
  artifact = NULL,
  repoUrl = getOption("rmaven.default_repos"),
  nocache = FALSE,
  verbose = c("normal", "quiet", "debug")
)
}
\arguments{
\item{groupId}{optional, the maven groupId,}

\item{artifactId}{optional, the maven artifactId,}

\item{version}{optional, the maven version,}

\item{...}{other maven coordinates such as classifier or packaging}

\item{coordinates}{optional, but if not supplied groupId and artifactId must be, coordinates as a coordinates object (see as.coordinates())}

\item{artifact}{optional, coordinates as an artifact string \verb{groupId:artifactId:version[:packaging[:classifier]]} string}

\item{repoUrl}{the URLs of the repositories to check (defaults to maven central)}

\item{nocache}{normally artifacts are only fetched if required, nocache forces fetching}

\item{verbose}{how much output from maven, one of "normal", "quiet", "debug"}
}
\value{
the .m2 path of the artifact
}
\description{
Fetch an artifact from a repository into the local .m2 cache
}
\examples{
fetch_artifact(artifact="com.google.guava:guava:31.1-jre")
fetch_artifact(coordinates = as.coordinates("org.junit.jupiter","junit-jupiter-api","5.9.0"))
}
