% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-interface.R
\name{print.coordinates}
\alias{print.coordinates}
\title{Prints a coordinates object}
\usage{
\method{print}{coordinates}(x, ...)
}
\arguments{
\item{x}{a maven coordinates object}

\item{...}{ignored}
}
\value{
nothing. for side effects.
}
\description{
Prints a coordinates object
}
\examples{
print(as.coordinates("org.junit.jupiter","junit-jupiter-api","4.13.2"))
}
