% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maven.R
\name{copy_artifact}
\alias{copy_artifact}
\title{Fetch an artifact from a repository to a local directory}
\usage{
copy_artifact(
  repoUrl = getOption("rmaven.default_repos"),
  ...,
  coordinates = as.coordinates(...),
  artifact = .artifact(coordinates),
  outputDirectory = .working_dir(artifact),
  verbose = c("normal", "quiet", "debug")
)
}
\arguments{
\item{repoUrl}{the URLs of the repositories to check (defaults to maven central & sonatype snaphots)}

\item{...}{can express the coordinates as groupId, artifactId and version strings, plus optionally packaging and coordinates}

\item{coordinates}{optional, coordinates as a coordinates object (see as.coordinates())}

\item{artifact}{optional, coordinates as an artifact string groupId:artifactId:version[:packaging\link{:classifier}] string}

\item{outputDirectory}{optional path, defaults to the rmaven cache directory}

\item{verbose}{how much output from maven, one of "normal", "quiet", "debug"}
}
\value{
the output of the system2 call. 0 on success.
}
\description{
Fetch an artifact from a repository to a local directory
}
\examples{
tmp = copy_artifact(artifact="io.github.terminological:r6-generator:main-SNAPSHOT:pom")
}
