% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maven.R
\name{execute_maven}
\alias{execute_maven}
\title{Executes a maven goal}
\usage{
execute_maven(
  goal,
  opts = c(),
  pom_path = NULL,
  quiet = .quietly(verbose),
  debug = .debug(verbose),
  verbose = c("normal", "debug", "quiet"),
  require_jdk = FALSE,
  ...
)
}
\arguments{
\item{goal}{the goal of the \code{mvn} command ( can be multiple ) e.g. \code{c("clean","compile")}}

\item{opts}{provided options in the form \code{c("-Doption1=value2","-Doption2=value2")}}

\item{pom_path}{optional. the path to a pom.xml file for goals that operate on one}

\item{quiet}{should output from maven be suppressed? (-q flag)}

\item{debug}{should output from maven be verbose? (-X flag)}

\item{verbose}{how much output from maven, one of "normal", "quiet", "debug"}

\item{require_jdk}{does the goal you are executing require a \code{JDK} (e.g. compilation)}

\item{...}{named parameters are passed to maven as options in the form \code{-Dname=value}}
}
\value{
nothing, invisibly
}
\description{
Maven goals may be executed with or without a pom.xml file. Some maven goals (e.g. compilation)
require the use of a \code{JDK}.
}
\examples{
\donttest{
# This code can take quite a while to run as has to
# download a lot of plugins, especially on first run
execute_maven("help:system")
}
}
