% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maven.R
\name{fetch_artifact}
\alias{fetch_artifact}
\title{Fetch an artifact from a repository into the local .m2 cache}
\usage{
fetch_artifact(
  repoUrl = getOption("rmaven.default_repos"),
  ...,
  coordinates = as.coordinates(...),
  artifact = .artifact(coordinates),
  verbose = c("normal", "quiet", "debug")
)
}
\arguments{
\item{repoUrl}{the URLs of the repositories to check (defaults to maven central)}

\item{...}{can express the coordinates as groupId, artifactId and version strings, plus optionally packaging and coordinates}

\item{coordinates}{optional, but if not supplied groupId and artifactId must be, coordinates as a coordinates object (see as.coordinates())}

\item{artifact}{optional, coordinates as an artifact string groupId:artifactId:version[:packaging\link{:classifier}] string}

\item{verbose}{how much output from maven, one of "normal", "quiet", "debug"}
}
\value{
the output of the system2 call. 0 on success.
}
\description{
Fetch an artifact from a repository into the local .m2 cache
}
\examples{
fetch_artifact(artifact="io.github.terminological:r6-generator:main-SNAPSHOT:pom")
}
